changed IK() to use int32_t instead of long.

+++ src/robots/arm3.cpp
@@ -16,7 +16,7 @@
  * @param y cartesian coordinate
  * @param motorStepArray a measure of each belt to that plotter position
  */
-void IK(const float *const axies, long *motorStepArray) {
+void IK(const float *const axies, int32_t *motorStepArray) {
   float x = axies[0];
   float y = axies[1];
   float z = axies[2];
